requirements.txt
numpy==1.26.4
Pillow==10.1.0
opencv-python==4.9.0.80
pytesseract==0.3.13